% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{Server for the Palmer Penguins Shiny App}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}
}
\value{
A server function
}
\description{
Server for the Palmer Penguins Shiny App
}
