% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_error_report.R
\name{create_error_report}
\alias{create_error_report}
\title{Create Error Report}
\usage{
create_error_report(filepath, error_message, params)
}
\arguments{
\item{filepath}{The path to create the error report}

\item{error_message}{The error message}

\item{params}{The parameters that were to be used}
}
\value{
NULL (creates file as a side effect)
}
\description{
Create Error Report
}
\keyword{internal}
