% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report_template.R
\name{create_report_template}
\alias{create_report_template}
\title{Create Report Template}
\usage{
create_report_template(filepath)
}
\arguments{
\item{filepath}{The path to create the template file}
}
\value{
NULL (creates file as a side effect)
}
\description{
Create Report Template
}
\keyword{internal}
