% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report_template_content.R
\name{create_report_template_content}
\alias{create_report_template_content}
\title{Create Report Template Content}
\usage{
create_report_template_content()
}
\value{
A character string containing the R Markdown template
}
\description{
Create Report Template Content
}
\keyword{internal}
