% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch the Palmer Penguins Shiny Application}
\usage{
launch_app()
}
\value{
A Shiny application object
}
\description{
This function launches the Shiny application for exploring and
visualizing the Palmer Penguins dataset.
}
\examples{
if (interactive()) {
  launch_app()
}
}
