% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_download_report.R
\name{mod_download_report_server}
\alias{mod_download_report_server}
\title{Download Report Module Server}
\usage{
mod_download_report_server(id, data, inputs)
}
\arguments{
\item{id}{The module identifier}

\item{data}{The Palmer Penguins dataset}

\item{inputs}{A reactive list of user inputs}
}
\value{
NULL (outputs are rendered to the UI)
}
\description{
Download Report Module Server
}
