% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_graph_output.R
\name{mod_graph_output_server}
\alias{mod_graph_output_server}
\title{Graph Output Module Server}
\usage{
mod_graph_output_server(id, data, inputs)
}
\arguments{
\item{id}{The module identifier}

\item{data}{The Palmer Penguins dataset}

\item{inputs}{A reactive list of user inputs}
}
\value{
NULL (outputs are rendered to the UI)
}
\description{
Graph Output Module Server
}
