% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_graph_output.R
\name{mod_graph_output_ui}
\alias{mod_graph_output_ui}
\title{Graph Output Module UI}
\usage{
mod_graph_output_ui(id)
}
\arguments{
\item{id}{The module identifier}
}
\value{
A UI definition
}
\description{
Graph Output Module UI
}
