% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_input.R
\name{mod_input_server}
\alias{mod_input_server}
\title{Input Module Server}
\usage{
mod_input_server(id, data)
}
\arguments{
\item{id}{The module identifier}

\item{data}{The Palmer Penguins dataset}
}
\value{
A reactive list of user inputs
}
\description{
Input Module Server
}
