% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_table_output.R
\name{mod_table_output_ui}
\alias{mod_table_output_ui}
\title{Table Output Module UI}
\usage{
mod_table_output_ui(id)
}
\arguments{
\item{id}{The module identifier}
}
\value{
A UI definition
}
\description{
Table Output Module UI
}
